[ADD] finance_kmitl_demo: post-budget finance demo data
Move the disbursement-flow demo (PR -> PA -> PO -> WA -> DR) out of
kmitl_demo into a new dedicated demo module and extend it with the
accounting tail that had no demo data: vendor bills (tang-nee),
payments + KTB bank payment export + WHT (lang-nee), and fixed assets
with depreciation.

kmitl_demo keeps master data and the purchase end-to-end flow; the
shared helpers it still needs are imported, not duplicated. Records
are left across a realistic mix of workflow states and each risky
step is wrapped defensively so the post_init hook never aborts the
install.
