+++ finance_kmitl_demo/__init__.py
@@ -1,0 +1,1 @@
+from .hooks import post_init
